requirements-manager.txt
# ESP32 Device Manager Backend Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
pyserial>=3.5
paho-mqtt>=1.6.0
amqtt>=0.11.0
Pillow>=10.0.0
esptool>=4.6.0
python-multipart>=0.0.6
